+++ tests/testthat/test-replacement-patterns-file-format.R
@@ -1,8 +1,8 @@
 test_that("replacementPatterns.csv has valid format", {
   lines <- readLines(system.file("csv", "replacementPatterns.csv", package = "SqlRender"))
   for (line in lines) {
-    line_trim <- stringr::str_replace_all(line, '".[^"]+"', "")
-    comma_cnt <- stringr::str_count(line_trim, ",")
+    lineTrim <- gsub('"[^"]*"', "", gsub('\\\\"', "", line))
+    comma_cnt <- nchar(lineTrim) - nchar(gsub(",", "", lineTrim))
     expect_equal(comma_cnt, 2, label = paste0("comma counts in string: ", line))
   }
 })